requirements.txt
numpy
pandas
wandb
pytorch
pytorch-lightning

git+git://github.com/KIT-IAI/pyWATTS@master#egg=pywatts